requirements.txt
django
moviepy
celery
gunicorn
django-tinymce4-lite